Refactor backend folder structure
