+++ packages/backend/src/models/Contact.js
@@ -1,0 +1,10 @@
+import mongoose from "mongoose";
+
+const contactSchema = new mongoose.Schema({
+  name: String,
+  email: String,
+  message: String,
+  createdAt: { type: Date, default: Date.now },
+});
+
+export const Contact = mongoose.model("Contact", contactSchema);
