+++ packages/backend/src/models/User.js
@@ -1,0 +1,9 @@
+import mongoose from "mongoose";
+
+const userSchema = new mongoose.Schema({
+  email: { type: String, required: true, unique: true },
+  username: { type: String, required: true, unique: true },
+  hashedPassword: { type: String, required: true },
+});
+
+export const User = mongoose.model("User", userSchema);
