+++ packages/backend/src/routes/authRoutes.js
@@ -1,0 +1,10 @@
+import express from "express";
+import { registerUser, loginUser } from "../controllers/authController.js";
+
+const router = express.Router();
+
+// Authentication Routes
+router.post("/signup", registerUser);
+router.post("/login", loginUser);
+
+export default router;
