+++ packages/backend/src/server.js
@@ -1,0 +1,39 @@
+import express from "express";
+import mongoose from "mongoose";
+import cors from "cors";
+import dotenv from "dotenv";
+import taskRoutes from "./routes/taskRoutes.js";
+import contactRoutes from "./routes/contactRoutes.js";
+import authRoutes from "./routes/authRoutes.js";
+
+dotenv.config();
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+const mongoURI = process.env.MONGO_URI;
+if (!mongoURI) {
+  console.error("MONGO_URI is not defined in .env");
+  process.exit(1);
+}
+
+const PORT = process.env.PORT || 5001;
+
+// Connect to MongoDB
+mongoose
+  .connect(mongoURI)
+  .then(() => console.log("Connected to MongoDB Atlas"))
+  .catch((err) => {
+    console.error("MongoDB connection error:", err);
+    process.exit(1);
+  });
+
+// Routes
+app.use("/tasks", taskRoutes);
+app.use("/contact", contactRoutes);
+
+// Authentication Routes
+app.use("/auth", authRoutes);
+
+app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
