+++ packages/frontend/src/Signup.jsx
@@ -25,7 +25,7 @@ const Signup = ({ setToken }) => {
     }
 
     try {
-      const response = await fetch(`${API_BASE_URL}/signup`, {
+      const response = await fetch(`${API_BASE_URL}/auth/signup`, {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify({ email, username, password }),
